acceptance,server: move TestDebugRemote to an acceptance test
Change COCKROACH_REMOTE_DEBUG to server.debug.remote cluster setting.

Fixes #14714
